requirements.txt
# Core runtime dependencies
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.1

# Testing & quality
pytest==8.3.2
coverage==7.6.1
flake8==7.1.1
bandit==1.7.9
safety==3.2.4

# (Optional) Add gunicorn/waitress for production WSGI serving if needed
